requirements.txt
aerich==0.5.5
aiosqlite==0.16.1
asgiref==3.4.1
asyncpg==0.23.0
click==8.0.1
ddlparse==1.10.0
decorator==5.0.9
dictdiffer==0.9.0
fastapi==0.68.0
h11==0.12.0
httptools==0.2.0
iso8601==0.1.16
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
starlette==0.14.2
tortoise-orm==0.17.6
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.3
validators==0.18.2
watchgod==0.7
websockets==9.1
